discord_bot/requirements.txt
discord.py
APScheduler
python-dotenv
fastapi
uvicorn
requests
